+++ examples/scripts/sft.py
@@ -81,7 +81,7 @@ if TRL_USE_RICH:
 
 
 if __name__ == "__main__":
-    parser = TrlParser(SftScriptArguments, TrainingArguments, ModelConfig)
+    parser = TrlParser((SftScriptArguments, TrainingArguments, ModelConfig))
     args, training_args, model_config = parser.parse_args_and_config()
 
     # Force use our print callback
